Loosen the version restrictions for dependencies
The best approach for libraries is to assume the newer major versions of dependencies are backwards-compatible to the extent that we user them, breaking only in rare edge cases, which we likely do not use.

If it breaks, the app developers can put additional restrictions on their side in their virtualenvs. It is NOT the job of a library to restrict the dependencies from the upper side.

Only the lower restrictions should be set — to the minimum version supporting specific functionality that we use. Again, as lower as possible — to let the app developers space for maneuvering within their virtualenvs and whatever legacy code they have, as long as it does not break our code.